Fixed Bugs with new HD writing approach
1) Fixed error that if processing occurs for RAM for say 4 levels then
all of a sudden we need to switch to HD.  We need to make sure all of
the pLists are ordered this only has to be done once if it has never
been done after the first level.  I may add a vector of bools to flag
this stuff.  But usually once a transition has been made from HD to RAM
then most of the processing thereafter is RAM because pattern counts are
getting smaller and smaller.
2) Created a static thread handler for the msync threads.  It
essentially deallocates these spawned threads.  F OO