RAM fix
1) Added a write flag to prevent exception when adding new chunks of the
processed file to the fileChunks vector.  Reading of the file is blocked
when a new chunk is being added.
2) Encountered an old bug in the code that had gone unnoticed in the
first level of processing for ram.  If we were not specifically saying
we are going to use ram and the prediction is seen as using ram then we
would still use file read size truncation.