Huge RAM speedup
 1) Cleaned up some warning code.
 2) Fixed memory issue.
 3) Found another bottleneck in constantly searching through the
treemap to find if a pattern exists in the map and if so add memory
count.  This can be done simply by just adding the size of a pList per
addition. Increased speed of processing from 280 to 270 seconds.  We
will have to see if this affects processing of the HD but it looks okay
for now.
 4) Added more const reference types to increase speed.   Added inline
for TreeHD addLeaf and AddPList and the speed change went from 120 to
110.
 5) Using unordered_map instead of map to test for speed increase
possibility.  Speed went from 153 to 120 seconds for the Boosh.avi.
This was the best improvement overall in this promote.
 6) Fixed the thread polling issue by calling valid() instead of
wait_for(milliseconds(5)).  I now call valid (i.e. join) and if the
thread is not ready then call thread::Sleep(milliseconds(5)).  Does the
same job except no crashes now.  Find a good amount of milliseconds to
wait is key because we don’t want to overload the system with threads
and we want to poll at a fast enough rate to put cores to work if we
have available threads.
 7) Unordered_map seems to be the best solution for TreeRAM and not for
TreeHD.  I believe the reason is that TreeHD can have very large strings
in them for keys and that slows searching in find.