Preload string information before processing
Program processes on linux in 11011 mS which is the fastest it has ever
processed Boosh.avi with 81 threads.  Another example is the same file
with 4 threads that gets a speed improvement of 87 to 70 seconds.
Preloading the string information causes a lot less cache misses because
we are not battling with the other data that needs to be in cache.
Excellent!