Fixed elimination pattern bug
Fixed a bug in the program that was causing pattern elimination tracking
to fail.
Fixed a bug in the second level of programming where there could be
multiple vectors which contain the same pattern that were built by the
first level algorithm.
This fix gave a boost in processing speed also.