Fixed Virtual Memory issue and improved speed!
Virtual Memory was building up and would eventual crash the program.
The fix involved querying the linux box what the virtual memory page
size was and then memory mapping to multiples of the queried page size.
Improved speed by not constantly opening and closing files.  If a file
were to be used again then the handle would be saved for later use.