Throughput Scaling CSV generation
Made some changes that fixed opening too many files problem.
For ram processing bench tests I placed the called to load in data
before starting the processing timer because we are not concerned about
how long it takes to load in a file from the Hard Drive...there may be a
bug in this promote.