Fixed first level hard drive processing issue
Added logging if a file can't be deleted.
Added more file handling checks and added readability to the errors.
Fixed in issue in hard drive processing when a file has just been
deleted and the same file needs to be regenerated.  The fix involved
changing the name of file but keeping the same data.
MATLAB processing displays the count of most common pattern and the
coverage of the file.